% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecbAvailability.R
\name{ecbAvailability}
\alias{ecbAvailability}
\title{ecbAvailability: See which datasets are available via the ECB SDW}
\usage{
ecbAvailability()
}
\value{
data.table with the available datasets.
}
\description{
See the available datasets at the ECB SDW
}
\author{
Janko Cizel
}

