% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecbDataset.R
\name{ecbDataset}
\alias{ecbDataset}
\title{ecbDataset: download entire dataset from the ECB SDW}
\usage{
ecbDataset(datasetId = "ECB_AME1")
}
\arguments{
\item{datasetId}{Dataset ID, typically in the following format:
EXT_[3-letter flowRef (see below)][#]}
}
\value{
Downloaded dataset. It contains a DSD attribute, which can be useful
    in labeling data series.
}
\description{
Download a specified dataset from the ECB SDW
}
\author{
Janko Cizel
}

