% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecbReshape.R
\name{ecbReshape}
\alias{ecbReshape}
\title{ecbReshape: Reshape ecbDataset or ecbSeries outputs into a wide format}
\usage{
ecbReshape(data, colId)
}
\arguments{
\item{data}{Input dataset, which must be an output of "ecbDataset" or "ecbSeries" functions!}

\item{colId}{Name of a dimension that is to be transformed into columns}
}
\value{
data.table in a wide format
}
\description{
Reshape ecbDataset or ecbSeries outputs into a wide format
}
\author{
Janko Cizel
}

