% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecbReshape2ts.R
\name{ecbReshape2ts}
\alias{ecbReshape2ts}
\title{ecbReshape2ts}
\usage{
ecbReshape2ts(data)
}
\arguments{
\item{data}{Input dataset, which must be an output of "ecbDataset" or "ecbSeries" functions!}
}
\value{
data.table in a wide format
}
\description{
Reshape the outputs of 'ecbDataset' or 'ecbSeries' functions into a
series-based wide format
}
\author{
Janko Cizel
}

