% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecbSeries.R
\name{ecbSeries}
\alias{ecbSeries}
\title{ecbSeries: download matching data series from the ECB SDW}
\usage{
ecbSeries(datasetId = "ECB_AME1", pattern = ".GBR.....ZUTN")
}
\arguments{
\item{datasetId}{Dataset ID, typically in the following format:
EXT_[3-letter flowRef (see below)][#]}

\item{pattern}{Pattern specifying the data series names to be downloaded}
}
\value{
Downloaded dataset. It contains a DSD attribute, which can be useful
    in labeling data series.
}
\description{
Download specified data series from the ECB SDW
}
\author{
Janko Cizel
}

