% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datastructure.R
\name{get_codelists}
\alias{get_codelists}
\title{Utility function to obtain codelists}
\usage{
get_codelists(lookup)
}
\arguments{
\item{lookup}{}
}
\description{
Utility function to obtain codelists
}
\author{
Janko Cizel
}

