% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\title{get_dataset: download entire dataset from ECB SDW}
\usage{
get_dataset(datasetId = "ECB_AME1", flowRef = substr(datasetId, 5, 7))
}
\arguments{
\item{datasetId}{Dataset ID, typically in the following format:
EXT_[3-letter flowRef (see below)][#]}

\item{flowRef}{Optional argument. By default it is assumed that flowRef is
contained in the datasetId code (see above).}
}
\value{
Downloaded dataset. It contains a DSD attribute, which can be useful
    in labeling data series.
}
\description{
Download entire dataset from ECB SDW.
}
\author{
Janko Cizel
}

