% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_datasets}
\alias{get_datasets}
\title{View all available datasets from a specified data source}
\usage{
get_datasets(source = "ECB")
}
\arguments{
\item{source}{Name of the source. The default is ECB.}
}
\value{
A data.table with information about the available datasets.
}
\description{
View all available datasets from a specified data source
}
\author{
Janko Cizel
}

