% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datastructure.R
\name{get_datastructure}
\alias{get_datastructure}
\title{Get SDMX datastructure of the dataset at ECB SDW}
\usage{
get_datastructure(source = "ECB", dataset = "ECB_AME1")
}
\arguments{
\item{url}{URL of SDMX datastructure (available at SDW)}
}
\value{
Returns a list of lookup tables for each of the dimensions in the
    data structure
}
\description{
get_datastructure: Get SDMX datastructure of the dataset
}
\author{
Janko Cizel
}

