% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multiple_series.R
\name{get_multiple_series}
\alias{get_multiple_series}
\title{Download multiple series}
\usage{
get_multiple_series(dataset, series_ids)
}
\arguments{
\item{dataset}{}

\item{series_ids}{}
}
\description{
Utility function to download multiple series
}
\details{
Download multiple series
}
\author{
Janko Cizel
}

