% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ameco.R
\name{get_series_ameco}
\alias{get_series_ameco}
\title{Get all available AMECO series}
\usage{
get_series_ameco(concept_id = "ZUTN")
}
\arguments{
\item{concept_id}{optionally, specify (part of) series code to look for}
}
\value{
Returns a character vector of matching series.
}
\author{
Janko Cizel
}

