% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_static.R
\name{get_static}
\alias{get_static}
\title{Get static information for the specified dataset.}
\usage{
get_static(source = "ECB", dataset = "ECB_IVF1")
}
\arguments{
\item{source}{Name of the data source. Defautl is the ECB.}

\item{dataset}{Name of the dataset.}
}
\value{
A data.table with static information.
}
\description{
Get static information for the specified dataset.
}
\author{
Janko Cizel
}

