% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\title{info: get further information on the dataset}
\usage{
info(data)
}
\arguments{
\item{data}{data Input dataset, which must be an output of "ecbDataset" or "ecbSeries" functions!}
}
\value{
data.table with info on the dataset dimensions
}
\description{
Get further information on the dataset
}
\author{
Janko Cizel
}

