% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Download an ECB dataset}
\usage{
get_data(dataset = "AME")
}
\arguments{
\item{dataset}{Name of the dataset to download.}
}
\value{
A data.table containing the requested data.
}
\description{
This function downloads an ECB dataset, specified by the user. The available
datasets can be viewed by using the `get_datasets` function.
}
\author{
Janko Cizel
}

